refactor: global structure
