+++ webAleaFront/src/features/logger/ui/ConsoleLog.jsx
