+++ webAleaFront/src/features/logger/ui/LogLine.jsx
