+++ webAleaFront/src/features/nodes/model/NodeInputBool.jsx
