+++ webAleaFront/src/features/nodes/model/NodeInputFloat.jsx
