+++ webAleaFront/src/features/nodes/model/NodeInputStr.jsx
