+++ webAleaFront/src/features/nodes/ui/NodeOutputs.jsx
