+++ webAleaFront/src/features/nodes/ui/sidebar_detail/NodeDescription.jsx
