+++ webAleaFront/src/features/nodes/ui/sidebar_detail/NodeDetailSection.jsx
