+++ webAleaFront/src/features/package-manager/ui/type/README.md
@@ -1,0 +1,2 @@
+### TODO
+Refactorisation des deux types. Ce sont des codes dupliqués.
