+++ webAleaFront/src/features/workspace/Workspace.jsx
