+++ webAleaFront/src/features/workspace/model/Node.jsx
