+++ webAleaFront/src/features/workspace/ui/CustomNode.jsx
