+++ webAleaFront/src/features/workspace/ui/type/StringNode.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Handle, Position } from "@xyflow/react";
 import "../../assets/css/custom_node.css";
-import { useLog } from "../../providers/LogContextDefinition.jsx";
+import { useLog } from "../../../logger/providers/LogContextDefinition.jsx";
 import { useFlow } from "../../providers/FlowContextDefinition";
 
 export default function StringNode(nodeProps) {
